+++ src/operator/nn/dropout-inl.h
@@ -256,7 +256,7 @@ class DropoutOp {
 
       // set dropout state.
       Random<xpu, unsigned> *prnd = ctx.requested[1].get_random<xpu, unsigned>(s);
-      uint64_t rng_seed = (uint64_t) prnd->GetSeed();
+      uint64_t rng_seed = prnd->GetSeed();
       // reset dropout descriptor if rng seed changed.
       bool reset = seed_ != rng_seed;
       seed_ = rng_seed;
